+++ man-roxygen/n_points.R
@@ -1,0 +1,1 @@
+#' @param n_points Number of points to resample the centerline to.